devide libft func from init_cubfile.c to src/utils
