+++ src/utils/ft_strlen.c
@@ -1,0 +1,12 @@
+#include "cubfile.h"
+#include <stddef.h>
+
+size_t	ft_strlen(const char *str)
+{
+	size_t	ret;
+
+	ret = 0;
+	while (str[ret])
+		++ret;
+	return (ret);
+}
